Separate modules, add findWithAssert helper
