+++ addon-test-support/trigger-event.js
@@ -1,0 +1,20 @@
+import Ember from 'ember';
+import { findWithAssert } from 'ember-native-dom-helpers/test-support/find-with-assert';
+import { fireEvent } from 'ember-native-dom-helpers/test-support/fire-event';
+import wait from 'ember-test-helpers/wait';
+
+const { run } = Ember;
+
+/*
+  @method triggerEvent
+  @param {String} selector
+  @param {String} type
+  @param {Object} options
+  @return {RSVP.Promise}
+  @public
+*/
+export function triggerEvent(selector, type, options) {
+  let element = findWithAssert(selector);
+  run(() => fireEvent(element, type, options));
+  return wait();
+}
